+++ src/main/java/gg/cozycrafters/cozydisplays/display/TextRenderMode.java
@@ -1,0 +1,6 @@
+package gg.cozycrafters.cozydisplays.display;
+
+public enum TextRenderMode {
+    LINE_ENTITIES,
+    SINGLE_ENTITY
+}